Initial Commit
Viewing the index and fourms is all that is done. They need to be
created manually using SQL statements.

Instead of CSS, I'm using Sass. It's much nicer, but we need to compile
it on runtime. Luckily, we only need to compile when the board starts,
and Sass caches everything, so it shouldn't affect much.

A proper install page and configuration file need to be added. These
are low priority, so they probably won't be added until later.