Keep users logged in using Sessions
Users can login using the login page and a value will be stored in a
Hash in the Board. This is checked for each page that a user views, and
we can test if a user is logged in by checking if their id is -1, since
that is the default value.

Lots of things were hidden from "visiors" (users who aren't logged in).
They can't create new Threads, nor make Posts. Login and Register
buttons were added to the outer template, which visitors can click on
any page.