Edit Profile Page that can update titles
A user can go to their profile page and they will be greeted with a new
link to /edit/profile/ will allow them to change different attributes
associated with their account. For now, only the title can be changed.

Users that are not logged in will be prompted to login if they go to
/edit/profile.